theme: refactor views
* merge both blueprints into a single blueprint again, as that's easier
  to handle from a naming perspective (e.g. for url_for(...))
* instead, split up the routes and filters into separate modules
