+++ invenio_app_rdm/theme/templates/semantic-ui/invenio_app_rdm/landing_page/details/side_bar.html
@@ -56,7 +56,7 @@
     {# dynamically create the list of export formats #}
     {%- for fmt, val in config.get("APP_RDM_RECORD_EXPORTERS", {}).items() -%}
       {%- set name = val.get("name", fmt) -%}
-      {%- set export_url = url_for('invenio_app_rdm_records_ui.export_record', pid_value=record.pid.pid_value, export_format=fmt) -%}
+      {%- set export_url = url_for('invenio_app_rdm.export_record', pid_value=record.pid.pid_value, export_format=fmt) -%}
       <a class="export-format" href="{{ export_url }}">{{ name }}</a>
     {%- endfor -%}
   </dd>
