+++ setup.py
@@ -96,10 +96,7 @@ setup(
             'invenio-app-rdm = invenio_app.cli:cli',
         ],
         'invenio_base.blueprints': [
-            ("invenio_app_rdm = invenio_app_rdm.theme.views:"
-             "ui_blueprint"),
-            ("invenio_app_rdm_records = invenio_app_rdm.theme.views:"
-             "records_ui_blueprint"),
+            "invenio_app_rdm = invenio_app_rdm.theme.views:ui_blueprint",
         ],
         'invenio_assets.webpack': [
             'invenio_app_rdm_theme = invenio_app_rdm.theme.webpack:theme',
